huge APPS change
previeously we had a rule that app file will never change.
it will play as a config file to open windows or run functions.
also a namespace for app developer. these concept aren't changed.
but we don't store the windows and previews in APPS[app.name] anymore.

it all are stored inside the app. which brings huge flexibilities like
anonymous apps can run a window without having a key on APPS.

!! the whole system wasn't built with support for anonymous apps.
we should make some changes in windowing and previews,
as we don't have an icon in the startBar for non-loaded apps.
we flagged the places that needs change with !!anmapp